simulation_engine/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
numpy>=1.26.0
scipy>=1.13.0
python-dotenv>=1.0.0
